requirements-vercel.txt
# Django and REST Framework
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
drf-spectacular>=0.27

# Database
psycopg2-binary>=2.9

# Environment
python-decouple>=3.8

# OpenAI
openai>=1.0

# Error tracking
sentry-sdk>=1.40

# Python utilities
python-multipart>=0.0.6
PyJWT>=2.8
